+++ tests/app/settings.py
@@ -80,3 +80,5 @@ TEMPLATES = [
         },
     },
 ]
+
+WAGTAILADMIN_BASE_URL = "http://localhost:8080"