+++ tests/app/settings.py
@@ -2,6 +2,8 @@
 
 import os
 
+from wagtail import VERSION as WAGTAIL_VERSION
+
 PROJECT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 BASE_DIR = os.path.dirname(PROJECT_DIR)
 
@@ -49,7 +51,7 @@ INSTALLED_APPS = [
     "wagtail.documents",
     "wagtail.images",
     "wagtail.admin",
-    "wagtail.core",
+    "wagtail" if WAGTAIL_VERSION >= (3, 0) else "wagtail.core",
 ]
 
 ROOT_URLCONF = "tests.app.urls"
@@ -80,3 +82,5 @@ TEMPLATES = [
         },
     },
 ]
+
+WAGTAILADMIN_BASE_URL = "http://localhost:8080"