+++ tests/test_managers.py
@@ -1,7 +1,13 @@
 from django.shortcuts import reverse
 from django.test import TestCase
-from wagtail.core.models import Page
-from wagtail.tests.utils import WagtailTestUtils
+from wagtail import VERSION as WAGTAIL_VERSION
+
+if WAGTAIL_VERSION >= (3, 0):
+    from wagtail.models import Page
+    from wagtail.test.utils import WagtailTestUtils
+else:
+    from wagtail.core.models import Page
+    from wagtail.tests.utils import WagtailTestUtils
 
 from tests.app.models import TestPage
 from wagtail_trash.models import TrashCan, TrashCanPage