+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ### Fixed
 - Add python 3.11 support (Martin Sandström)
 
+### Removed
+- Drop Wagtail 2.16 support
+- Drop Wagtail 2.14 support
+
 ## 0.3.0
 - Add Wagtail 3.0 compatibility (Thanks @polesello!)
 - Add code and tests verifying urls gets preserved when moving pages (Andreas Bernacca)