+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Fixed
 - Add python 3.11 support (@marteinn)
 - Remove wagtailadmin.W003 warning in development (@marteinn)
+- Fix incorrect test alias in docker-entrypoint.sh (@marteinn)
 
 ### Removed
 - Drop Wagtail 2.16 support